backend/app/pyproject.toml
[project]
name = "adbudget-api"
version = "0.1.0"
description = "A mini-marketing analysis API that let users create ad campaigns and track speniture."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.41",
    "strawberry-graphql[fastapi]>=0.278.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
